+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_request_handlers.py
@@ -14,7 +14,7 @@ from ._models import (
     RecognizePiiEntitiesAction,
     RecognizeLinkedEntitiesAction,
     AnalyzeSentimentAction,
-    AnalyzeActionsType,
+    _AnalyzeActionsType,
 )
 
 def _validate_input(documents, hint, whole_input_hint):
@@ -71,14 +71,14 @@ def _validate_input(documents, hint, whole_input_hint):
 
 def _determine_action_type(action):
     if isinstance(action, RecognizeEntitiesAction):
-        return AnalyzeActionsType.RECOGNIZE_ENTITIES
+        return _AnalyzeActionsType.RECOGNIZE_ENTITIES
     if isinstance(action, RecognizePiiEntitiesAction):
-        return AnalyzeActionsType.RECOGNIZE_PII_ENTITIES
+        return _AnalyzeActionsType.RECOGNIZE_PII_ENTITIES
     if isinstance(action, RecognizeLinkedEntitiesAction):
-        return AnalyzeActionsType.RECOGNIZE_LINKED_ENTITIES
+        return _AnalyzeActionsType.RECOGNIZE_LINKED_ENTITIES
     if isinstance(action, AnalyzeSentimentAction):
-        return AnalyzeActionsType.ANALYZE_SENTIMENT
-    return AnalyzeActionsType.EXTRACT_KEY_PHRASES
+        return _AnalyzeActionsType.ANALYZE_SENTIMENT
+    return _AnalyzeActionsType.EXTRACT_KEY_PHRASES
 
 def _check_string_index_type_arg(string_index_type_arg, api_version, string_index_type_default="UnicodeCodePoint"):
     string_index_type = None